create APIRequest DTO
