+++ ChatBot/ChatBot/Models/Message.swift
@@ -1,0 +1,13 @@
+//
+//  Message.swift
+//  ChatBot
+//
+//  Created by Janine on 1/4/24.
+//
+
+import Foundation
+
+struct Message: Codable {
+    let role: String
+    let content: String
+}
